Optimiization

+++ include/celestial_body.hpp
@@ -4,14 +4,18 @@
 #include <vector>
 #include "SFML/Graphics.hpp"
 
+/**
+ * @brief Celestial body struct that contains the mass, radius, position, previous position, velocity and acceleration of a celestial body.
+ * @details total size of struct = 4*2 + 8*4 = 40 bytes.
+ */
 struct Celestial_body
 {   
-    double mass;
-    float radius;
-    sf::Vector2f position;
-    sf::Vector2f prev_position;
-    sf::Vector2f velocity;
-    sf::Vector2f acceleration;
+    float mass;                 // 4 bytes
+    float radius;               // 4 bytes
+    sf::Vector2f position;      // 8 bytes
+    sf::Vector2f prev_position; // 8 bytes
+    sf::Vector2f velocity;      // 8 bytes
+    sf::Vector2f acceleration;  // 8 bytes
 };
 
 #endif // CELESTIAL_BODY_HPP
